Remove EventHubClient altogether
